+++ hazelcast/src/main/java/com/hazelcast/cardinality/impl/CardinalityEstimatorService.java
@@ -219,7 +219,7 @@ public class CardinalityEstimatorService
 
     private class Merger implements Runnable {
 
-        private static final int TIMEOUT_FACTOR = 500;
+        private static final long TIMEOUT_FACTOR = 500;
 
         private final ILogger logger = nodeEngine.getLogger(CardinalityEstimatorService.class);
         private final Semaphore semaphore = new Semaphore(0);
